+++ src/extension.ts
@@ -4,32 +4,34 @@ import * as vscode from "vscode";
 import { opcodes, opcodesWithAddr, opcodeNames } from "./opcodes";
 
 function getAllLabels(document: vscode.TextDocument) {
-    const labels: [string, string][] = [];
+    const labels: Map<string, number> = new Map();
     const { locn, label, op } = colIndices(document);
-    const lines = document
-        .getText()
-        .split("\n")
-        .slice(1);
-    for (const line of lines) {
-        const labelTx = line.substring(label, op).split(" ")[0];
+    const lines = document.getText().split("\n");
+    for (
+        let i = +document.lineAt(0).text.startsWith("mem");
+        i < lines.length;
+        i++
+    ) {
+        const line = lines[i];
+        const labelTx = line.substring(label, op).trim();
         if (labelTx.length !== 0) {
-            labels.push([labelTx, line.substr(locn, 3)]);
+            labels.set(labelTx, i);
         }
     }
     return labels;
 }
 
-function findLabel(document: vscode.TextDocument, addr: string): string {
-    const lines = document.getText().split("\n");
-    const { locn, label, op } = colIndices(document);
-    for (const line of lines) {
-        if (line.indexOf(addr) === locn) {
-            return line.substring(label, op).split(" ")[0];
-        }
-    }
-    console.log("No match of label is found");
-    return "";
-}
+// function findLabel(document: vscode.TextDocument, addr: string): string {
+//     const lines = document.getText().split("\n");
+//     const { locn, label, op } = colIndices(document);
+//     for (const line of lines) {
+//         if (line.indexOf(addr) === locn) {
+//             return line.substring(label, op).trim();
+//         }
+//     }
+//     console.log("No match of label is found");
+//     return "";
+// }
 
 /**
  * get the indices of each column
@@ -38,14 +40,36 @@ function findLabel(document: vscode.TextDocument, addr: string): string {
  */
 function colIndices(document: vscode.TextDocument) {
     const line = document.lineAt(0).text;
-    return {
-        mem: line.indexOf("mem"),
-        locn: line.indexOf("locn"),
-        label: line.indexOf("label"),
-        op: line.indexOf("op"),
-        addr: line.indexOf("addr"),
-        comments: line.indexOf("comments")
-    };
+    if (line.startsWith("mem")) {
+        return {
+            mem: line.indexOf("mem"),
+            locn: line.indexOf("locn"),
+            label: line.indexOf("label"),
+            op: line.indexOf("op"),
+            addr: line.indexOf("addr"),
+            comments: line.indexOf("comments")
+        };
+    } else {
+        return {
+            mem: 0,
+            locn: 5,
+            label: 11,
+            op: 18,
+            addr: 27,
+            comments: 34
+        };
+    }
+}
+
+/**
+ * Convert zero-based line number to a IBCM memory address
+ * @param lineNum
+ */
+function getLocn(lineNum: number) {
+    return (
+        "0".repeat(+(lineNum < 16) + +(lineNum < 256)) +
+        lineNum.toString(16).toUpperCase()
+    );
 }
 
 interface IBCMLine {
@@ -92,6 +116,33 @@ export function activate(context: vscode.ExtensionContext) {
     // This line of code will only be executed once when your extension is activated
     console.log('Congratulations, your extension "ibcm" is now active!');
 
+    let fixCommand = vscode.commands.registerTextEditorCommand(
+        "ibcm.fixLocn",
+        (editor, edit, c) => {
+            const hasHeading = +editor.document
+                .lineAt(0)
+                .text.startsWith("mem");
+            const lines = editor.document.getText().split("\n");
+            const { locn } = colIndices(editor.document);
+            for (let lineNum = hasHeading; lineNum < lines.length; lineNum++) {
+                const line = lines[lineNum];
+                const originalLocn = line.substr(locn, 3);
+                const actualLocn = getLocn(lineNum - hasHeading);
+                if (line.length !== 0 && originalLocn !== actualLocn) {
+                    edit.replace(
+                        new vscode.Range(
+                            new vscode.Position(lineNum, locn),
+                            new vscode.Position(lineNum, locn + 3)
+                        ),
+                        actualLocn
+                    );
+                }
+            }
+        }
+    );
+
+    context.subscriptions.push(fixCommand);
+
     context.subscriptions.push(
         vscode.languages.registerHoverProvider("ibcm", {
             provideHover(document, position, token) {
@@ -105,15 +156,19 @@ export function activate(context: vscode.ExtensionContext) {
 
                     const oprand = code.substr(1);
                     const opcode = opcodes[parseInt(code.charAt(0), 16)];
+                    const indices = colIndices(document);
+                    const hasHeading = +document
+                        .lineAt(0)
+                        .text.startsWith("mem");
 
                     const desc = new vscode.MarkdownString();
                     if (opcodesWithAddr.has(opcode.op)) {
                         const targetLine = document.lineAt(
-                            parseInt(oprand, 16) + 1
+                            parseInt(oprand, 16) + hasHeading
                         ).text;
                         const label = targetLine
-                            .substr(colIndices(document).label)
-                            .split(" ")[0];
+                            .substring(indices.label, indices.op)
+                            .trim();
                         if (label.length !== 0) {
                             desc.appendMarkdown(
                                 opcode.desc(`**${label}**: ${oprand}`)
@@ -149,17 +204,14 @@ export function activate(context: vscode.ExtensionContext) {
         "ibcm",
         {
             provideCompletionItems(document, position, token) {
+                const hasHeading = +document.lineAt(0).text.startsWith("mem");
                 const indices = colIndices(document);
                 const { comments, addr } = indices;
                 const parts = document
                     .lineAt(position.line)
-                    .text.substring(0, position.character + 1)
+                    .text.substring(0, position.character + hasHeading)
                     .split(".");
-                const lineNum = position.line - 1;
-                const locnStr =
-                    "0".repeat(+(lineNum < 16) + +(lineNum < 256)) +
-                    lineNum.toString(16).toUpperCase();
-
+                const locnStr = getLocn(position.line - hasHeading);
                 if (parts.length === 1) {
                     const completionItems: vscode.CompletionItem[] = [];
                     for (const opcode of opcodes) {
@@ -181,6 +233,7 @@ export function activate(context: vscode.ExtensionContext) {
                                 new vscode.Position(position.line, comments)
                             );
                         }
+                        item.detail = opcode.desc("[mem]");
                         completionItems.push(item);
                     }
                     // code snippet for `dw` -- variable definition
@@ -219,9 +272,9 @@ export function activate(context: vscode.ExtensionContext) {
                         const completionItems: vscode.CompletionItem[] = [];
                         const options = [
                             ["1000", "readH", "read hex from keyboard"],
-                            ["1100", "readA", "read ASCII from keyboard"],
-                            ["1200", "writeH", "print hex to screen"],
-                            ["1300", "writeA", "print ASCII to screen"]
+                            ["1400", "readA", "read ASCII from keyboard"],
+                            ["1800", "writeH", "print hex to screen"],
+                            ["1C00", "writeA", "print ASCII to screen"]
                         ];
 
                         for (const [inst, itemLabel, comment] of options) {
@@ -250,9 +303,9 @@ export function activate(context: vscode.ExtensionContext) {
                         const completionItems: vscode.CompletionItem[] = [];
                         const options = [
                             ["200${1}", "shiftL", "shift left by ${1} bits"],
-                            ["210${1}", "shiftR", "shift right by ${1} bits"],
-                            ["220${1}", "rotL", "rotate left by ${1} bits"],
-                            ["230${1}", "rotR", "rotate right by ${1} bits"]
+                            ["240${1}", "shiftR", "shift right by ${1} bits"],
+                            ["280${1}", "rotL", "rotate left by ${1} bits"],
+                            ["2C0${1}", "rotR", "rotate right by ${1} bits"]
                         ];
 
                         for (const [inst, itemLabel, comment] of options) {
@@ -282,11 +335,14 @@ export function activate(context: vscode.ExtensionContext) {
                         // completion for opcode with an address, e.g. store.a
                         const completionItems: vscode.CompletionItem[] = [];
                         for (const temp of getAllLabels(document)) {
-                            const [label, locn] = temp;
+                            const [label, lineNum] = temp;
                             const item = new vscode.CompletionItem(
                                 label,
                                 vscode.CompletionItemKind.Variable
                             );
+                            const locn = document
+                                .lineAt(lineNum)
+                                .text.substr(indices.locn, 3);
                             item.detail = opcode.desc(`${label}: ${locn}`);
                             item.filterText = opcodeName + "." + label;
 
@@ -296,18 +352,20 @@ export function activate(context: vscode.ExtensionContext) {
 
                             item.documentation.appendCodeblock(
                                 document
-                                    .lineAt(parseInt(locn, 16) + 1)
+                                    .lineAt(lineNum)
                                     .text.substring(0, addr),
                                 "ibcm"
                             );
 
                             const line = addLine(document, {
-                                opcode: opcode.op.toString(16) + locn,
+                                opcode:
+                                    opcode.op.toString(16).toUpperCase() + locn,
                                 locn: locnStr,
                                 op: opcode.name,
                                 addr: label,
-                                comments: opcode.desc(`${label}:${locn}`)
+                                comments: opcode.desc(`\`${label}\``)
                             });
+
                             item.insertText = line;
                             item.range = new vscode.Range(
                                 new vscode.Position(position.line, 0),
@@ -324,6 +382,81 @@ export function activate(context: vscode.ExtensionContext) {
     );
 
     context.subscriptions.push(completionProvider);
+
+    const ibcmDiag = vscode.languages.createDiagnosticCollection("ibcm");
+    context.subscriptions.push(ibcmDiag);
+
+    const listener = vscode.workspace.onDidChangeTextDocument(editor => {
+        console.log("Editor changed");
+        ibcmDiag.clear();
+        const document = editor.document;
+        const indices = colIndices(document);
+        const lines = document.getText().split("\n");
+        const labels = getAllLabels(document);
+        const diags: vscode.Diagnostic[] = [];
+        const hasHeading = +document.lineAt(0).text.startsWith("mem");
+        for (let i = hasHeading; i < lines.length; i++) {
+            const line = lines[i];
+            const opcode = line.substr(0, 4);
+            const op = parseInt(opcode.charAt(0), 16);
+            // don't bother with instructions without a memory address
+            if (!opcodesWithAddr.has(op)) continue;
+            if (opcode.length !== 0) {
+                // the label of which this line of code refers as
+                const referredLabel = line
+                    .substring(indices.addr, indices.comments)
+                    .trim();
+                // const refereeLocn = labels.get(referee);
+                if (referredLabel.length !== 0) {
+                    const refereeLine =
+                        parseInt(opcode.substr(1), 16) + hasHeading;
+                    const actualLabel = document
+                        .lineAt(refereeLine)
+                        .text.substring(indices.label, indices.op)
+                        .trim();
+                    const refereeRange = new vscode.Range(
+                        new vscode.Position(refereeLine, indices.label),
+                        new vscode.Position(refereeLine, indices.op - 1)
+                    );
+                    const opcodeRange = new vscode.Range(
+                        new vscode.Position(i, 0),
+                        new vscode.Position(i, 3)
+                    );
+                    // empty target label
+                    if (actualLabel.length === 0) {
+                        const diag = new vscode.Diagnostic(
+                            refereeRange,
+                            `Warning: Label referred as ${referredLabel}, but is missing a label`,
+                            vscode.DiagnosticSeverity.Warning
+                        );
+                        diag.relatedInformation = [
+                            new vscode.DiagnosticRelatedInformation(
+                                new vscode.Location(document.uri, opcodeRange),
+                                `Referred by instruction ${opcode}`
+                            )
+                        ];
+                        diags.push(diag);
+                        // label mismatch
+                    } else if (actualLabel !== referredLabel) {
+                        const diag = new vscode.Diagnostic(
+                            refereeRange,
+                            `Warning: Label referred as ${referredLabel}, but is actually ${actualLabel}`,
+                            vscode.DiagnosticSeverity.Warning
+                        );
+                        diag.relatedInformation = [
+                            new vscode.DiagnosticRelatedInformation(
+                                new vscode.Location(document.uri, opcodeRange),
+                                `Referred by instruction ${opcode}`
+                            )
+                        ];
+                        diags.push(diag);
+                    }
+                }
+            }
+        }
+        ibcmDiag.set(document.uri, diags);
+    });
+    context.subscriptions.push(listener);
 }
 
 // this method is called when your extension is deactivated